compiler fix

+++ circuits/cpp/barretenberg/cpp/src/barretenberg/common/container.hpp
@@ -53,7 +53,7 @@ InnerCont flatten(Cont<InnerCont, Args...> const& in)
 
 // Return the first index at which a given item can be found in the vector.
 // Only safe for vectors with length less than the size_t overflow size.
-template <typename T> size_t index_of(std::vector<T> const& vec, T const& item)
+template <typename T> int64_t index_of(std::vector<T> const& vec, T const& item)
 {
     auto const& begin = vec.begin();
     auto const& end = vec.end();
